refactor: Use tGlobal in OperatorActionButton
